Create basic class for parser
