+++ .gitignore
@@ -1,0 +1,6 @@
+# pycache
+__pycache__/
+
+# Temporary files
+
+*.swp
